CHOUETTE-2483 Add required resources to display "Define Journey Pattern distances from Shape" macro

+++ app/models/macro/compute_journey_pattern_distances.rb
@@ -109,7 +109,7 @@ module Macro
                 SELECT
                   stop_area_pair,
                   journey_pattern_id AS journey_pattern_id,
-                  avg(distance) AS avg_distance
+                  avg(distance)::integer AS avg_distance
                 FROM (
                   SELECT
                     departure.journey_pattern_id AS journey_pattern_id,
